+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/repositories/MovieRepository.kt
@@ -4,6 +4,6 @@ import com.denisyordanp.mymoviecatalogue.schemas.ui.Movie
 import kotlinx.coroutines.flow.Flow
 
 interface MovieRepository {
-    suspend fun reloadMovies(genreId: Int)
-    fun getMovies(genreId: Int): Flow<List<Movie>>
+    suspend fun reloadMovies(genreId: Long)
+    fun getMovies(genreId: Long): Flow<List<Movie>>
 }
