+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/schemas/ui/MovieDetail.kt
@@ -1,16 +1,16 @@
 package com.denisyordanp.mymoviecatalogue.schemas.ui
 
 data class MovieDetail(
-    val id: Int,
+    val id: Long,
     val backdropPath: String,
-    val budget: Int,
+    val budget: Long,
     val genres: List<Genre>,
     val overview: String,
     val posterPath: String,
     val releaseDate: String,
-    val revenue: Int,
+    val revenue: Long,
     val tagline: String,
     val title: String,
     val voteAverage: Double,
-    val voteCount: Int
+    val voteCount: Long
 )
