+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/usecase/GetReviews.kt
@@ -4,5 +4,5 @@ import com.denisyordanp.mymoviecatalogue.schemas.ui.Review
 import kotlinx.coroutines.flow.Flow
 
 fun interface GetReviews {
-    operator fun invoke(movieId: Int, isForce: Boolean): Flow<List<Review>>
+    operator fun invoke(movieId: Long, isForce: Boolean): Flow<List<Review>>
 }
