+++ app/src/main/java/com/denisyordanp/mymoviecatalogue/usecase/GetVideos.kt
@@ -4,5 +4,5 @@ import com.denisyordanp.mymoviecatalogue.schemas.ui.Video
 import kotlinx.coroutines.flow.Flow
 
 fun interface GetVideos {
-    operator fun invoke(movieId: Int, isForce: Boolean): Flow<List<Video>>
+    operator fun invoke(movieId: Long, isForce: Boolean): Flow<List<Video>>
 }
